Fixed all of the issues in the targeting engine builder and made it generalize.

+++ src/main/java/com/zuehlke/pgadmissions/domain/imported/ImportedProgramSubjectArea.java
@@ -1,5 +1,7 @@
 package com.zuehlke.pgadmissions.domain.imported;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -29,7 +31,7 @@ public class ImportedProgramSubjectArea extends WeightedRelation implements Uniq
     private ImportedSubjectArea subjectArea;
 
     @Column(name = "relation_strength", nullable = false)
-    private Integer relationStrength;
+    private BigDecimal relationStrength;
 
     @Column(name = "enabled", nullable = false)
     private Boolean enabled;
@@ -61,12 +63,12 @@ public class ImportedProgramSubjectArea extends WeightedRelation implements Uniq
     }
 
     @Override
-    public Integer getRelationStrength() {
+    public BigDecimal getRelationStrength() {
         return relationStrength;
     }
 
     @Override
-    public void setRelationStrength(Integer relationStrength) {
+    public void setRelationStrength(BigDecimal relationStrength) {
         this.relationStrength = relationStrength;
     }
