+++ src/main/java/com/zuehlke/pgadmissions/domain/user/UserInstitution.java
@@ -1,5 +1,7 @@
 package com.zuehlke.pgadmissions.domain.user;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -28,7 +30,7 @@ public class UserInstitution extends UserImportedEntityRelation {
     private ImportedInstitution institution;
 
     @Column(name = "relation_strength", nullable = false)
-    private Integer relationStrength;
+    private BigDecimal relationStrength;
 
     @Override
     public Integer getId() {
@@ -59,12 +61,12 @@ public class UserInstitution extends UserImportedEntityRelation {
     }
 
     @Override
-    public Integer getRelationStrength() {
+    public BigDecimal getRelationStrength() {
         return relationStrength;
     }
 
     @Override
-    public void setRelationStrength(Integer relationStrength) {
+    public void setRelationStrength(BigDecimal relationStrength) {
         this.relationStrength = relationStrength;
     }
